refactor(ai): relocate browser-side fleet client to src/ai/fleet per Brain/Body separation (#14509)
The browser App-Worker fleet modules (fleetWireMethods, createFleetRegistryBridge, installFleetBridge) move from the Node Brain tree (ai/services/fleet/) to src/ai/fleet/ so the agentos app can import them: apps import from src/ (the Body/framework), never from ai/ (the Brain) — verified no app has a root-ai import. The Node transport (FleetControlBridge, dispatchFleetRequest, fleetBridgeServer) stays in ai/services/fleet/; dispatchFleetRequest imports the shared FLEET_WIRE_METHODS SSOT from src/ai/fleet/ (Brain-uses-framework is allowed; Body-uses-Brain is not).

29 specs green after the move.
